Ray tracing foveated working with MMF struct
The RenderMMF.h file contains all implementations of MMF method that render
the graphic image. This approach was development using the MMF struct as screen
to be painted, in other words, was defined a window (W) that was painted with
pixels with rays spaced by ( S / W ) pixels and between rays was used the color
of pixel calculated.