Graphic rendering foveated
In this update was considered the graphic rendering in domain of image.
Image was used by parameter and it's been modified at each iteration of levels.
- Remember that this project build a graphic scene using the MMF structure and
each pixel, in level, of MMF accesses a portion of scene to trace rays.
Was defined that only 1 ray would be analyzed by region and between regions,
the color of the previous calculated pixel will be used to fill this region.